@@ -1,6 +1,6 @@
 {
   "name": "@facetdev/node-cqrs-lib",
-  "version": "0.9.2",
+  "version": "0.9.3",
   "description": "Library of base classes for Node CQRS",
   "main": "index.js",
   "scripts": {
@@ -20,49 +20,49 @@
   },
   "homepage": "https://github.com/AttackPattern/node-cqrs-lib#readme",
   "dependencies": {
-    "async": "2.6.0",
-    "node-fetch": "1.7.3",
-    "uuid": "3.2.1"
+    "async": "2.6.1",
+    "node-fetch": "2.3.0",
+    "uuid": "3.3.2"
   },
   "devDependencies": {
-    "@babel/cli": "7.0.0",
-    "@babel/core": "7.0.0",
-    "@babel/plugin-proposal-object-rest-spread": "7.0.0",
-    "@babel/plugin-proposal-class-properties": "7.0.0",
-    "@babel/plugin-proposal-decorators": "7.0.0",
-    "@babel/plugin-proposal-optional-chaining": "7.0.0",
-    "@babel/plugin-proposal-export-default-from": "7.0.0",
-    "@babel/plugin-proposal-export-namespace-from": "7.0.0",
+    "@babel/cli": "7.2.0",
+    "@babel/core": "7.2.0",
+    "@babel/plugin-proposal-class-properties": "7.2.1",
+    "@babel/plugin-proposal-decorators": "7.2.0",
+    "@babel/plugin-proposal-export-default-from": "7.2.0",
+    "@babel/plugin-proposal-export-namespace-from": "7.2.0",
+    "@babel/plugin-proposal-object-rest-spread": "7.2.0",
+    "@babel/plugin-proposal-optional-chaining": "7.2.0",
     "@babel/polyfill": "7.0.0",
-    "@babel/preset-env": "7.0.0",
+    "@babel/preset-env": "7.2.0",
     "@babel/register": "7.0.0",
-    "@babel/runtime": "7.0.0",
-    "babel-eslint": "8.2.1",
-    "chai": "4.1.2",
+    "@babel/runtime": "7.2.0",
+    "babel-eslint": "10.0.1",
+    "chai": "4.2.0",
     "chai-as-promised": "7.1.1",
     "chai-datetime": "1.5.0",
     "del": "3.0.0",
-    "eslint": "4.15.0",
+    "eslint": "5.9.0",
     "gulp": "4.0.0",
-    "gulp-babel": "8.0.0-beta.2",
+    "gulp-babel": "8.0.0",
     "gulp-cached": "1.1.1",
     "gulp-changed": "3.2.0",
-    "gulp-debug": "3.2.0",
-    "gulp-eslint": "4.0.1",
+    "gulp-debug": "4.0.0",
+    "gulp-eslint": "5.0.0",
     "gulp-if": "2.0.2",
     "gulp-istanbul": "1.1.3",
     "gulp-load-plugins": "1.5.0",
-    "gulp-mocha": "5.0.0",
+    "gulp-mocha": "6.0.0",
     "gulp-remember": "1.0.1",
     "gulp-size": "3.0.0",
-    "gulp-sourcemaps": "2.6.3",
-    "gulp-uglify": "3.0.0",
+    "gulp-sourcemaps": "2.6.4",
+    "gulp-uglify": "3.0.1",
     "gulp-util": "3.0.8",
     "jshint-stylish": "2.2.1",
-    "mocha": "5.0.0",
-    "node-mocks-http": "1.6.6",
-    "proxyquire": "1.8.0",
-    "sinon": "4.1.6",
-    "supertest": "3.0.0"
+    "mocha": "5.2.0",
+    "node-mocks-http": "1.7.3",
+    "proxyquire": "2.1.0",
+    "sinon": "7.1.1",
+    "supertest": "3.3.0"
   }
 }
