@@ -1,2 +1,2 @@
-"use strict";Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=void 0;var _v=_interopRequireDefault(require("uuid/v4"));function _interopRequireDefault(e){return e&&e.__esModule?e:{default:e}}function _objectSpread(r){for(var e=1;e<arguments.length;e++){var t=null!=arguments[e]?arguments[e]:{},n=Object.keys(t);"function"==typeof Object.getOwnPropertySymbols&&(n=n.concat(Object.getOwnPropertySymbols(t).filter(function(e){return Object.getOwnPropertyDescriptor(t,e).enumerable}))),n.forEach(function(e){_defineProperty(r,e,t[e])})}return r}function asyncGeneratorStep(e,r,t,n,o,a,u){try{var s=e[a](u),i=s.value}catch(e){return void t(e)}s.done?r(i):Promise.resolve(i).then(n,o)}function _asyncToGenerator(s){return function(){var e=this,u=arguments;return new Promise(function(r,t){var n=s.apply(e,u);function o(e){asyncGeneratorStep(n,r,t,o,a,"next",e)}function a(e){asyncGeneratorStep(n,r,t,o,a,"throw",e)}o(void 0)})}}function _classCallCheck(e,r){if(!(e instanceof r))throw new TypeError("Cannot call a class as a function")}function _defineProperties(e,r){for(var t=0;t<r.length;t++){var n=r[t];n.enumerable=n.enumerable||!1,n.configurable=!0,"value"in n&&(n.writable=!0),Object.defineProperty(e,n.key,n)}}function _createClass(e,r,t){return r&&_defineProperties(e.prototype,r),t&&_defineProperties(e,t),e}function _defineProperty(e,r,t){return r in e?Object.defineProperty(e,r,{value:t,enumerable:!0,configurable:!0,writable:!0}):e[r]=t,e}var Repository=function(){function u(e){var s=this,r=e.eventStore,t=e.constructor,n=e.aggregateType,o=e.snapshots,a=void 0===o||o;_classCallCheck(this,u),_defineProperty(this,"get",function(){var r=_asyncToGenerator(regeneratorRuntime.mark(function e(r){var t,n,o,a,u;return regeneratorRuntime.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,s.eventStore.getEvents({aggregateType:s.aggregateType,aggregateId:r});case 2:if(n=e.sent,o=n.events,(a=n.snapshot)||o.length){e.next=7;break}return e.abrupt("return",null);case 7:if(u=new s.Constructor({id:r,events:o,snapshot:null==a?void 0:a.body,version:null==a?void 0:a.version}),!(s.snapshots&&((null===(t=u.$snapshotSchedule)||void 0===t?void 0:t.every)||Number.MAX_VALUE)<=o.length)){e.next=11;break}return e.next=11,s.eventStore.saveSnapshot(s.aggregateType,u);case 11:return e.abrupt("return",u);case 12:case"end":return e.stop()}},e,this)}));return function(e){return r.apply(this,arguments)}}()),_defineProperty(this,"create",function(e){return new s.Constructor({id:e||(0,_v.default)()})}),_defineProperty(this,"record",function(){var r=_asyncToGenerator(regeneratorRuntime.mark(function e(r){var t;return regeneratorRuntime.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,s.eventStore.record(r.map(function(e){return _objectSpread({},e,{aggregate:s.aggregateType})}));case 2:return t=e.sent,e.next=5,s.notifySubscribers(t);case 5:case"end":return e.stop()}},e,this)}));return function(e){return r.apply(this,arguments)}}()),_defineProperty(this,"notifySubscribers",function(){var r=_asyncToGenerator(regeneratorRuntime.mark(function e(t){return regeneratorRuntime.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,Promise.all(s.subscriptions.map(function(){var r=_asyncToGenerator(regeneratorRuntime.mark(function e(r){return regeneratorRuntime.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return e.prev=0,e.next=3,r(t);case 3:return e.abrupt("return",e.sent);case 6:return e.prev=6,e.t0=e.catch(0),console.log(e.t0),e.abrupt("return",null);case 10:case"end":return e.stop()}},e,this,[[0,6]])}));return function(e){return r.apply(this,arguments)}}()));case 2:return e.abrupt("return",e.sent);case 3:case"end":return e.stop()}},e,this)}));return function(e){return r.apply(this,arguments)}}()),_defineProperty(this,"getEvents",function(){var r=_asyncToGenerator(regeneratorRuntime.mark(function e(r){return regeneratorRuntime.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,s.eventStore.getEvents({aggregateType:s.aggregateType,aggregateId:r});case 2:return e.abrupt("return",e.sent);case 3:case"end":return e.stop()}},e,this)}));return function(e){return r.apply(this,arguments)}}()),this.eventStore=r,this.aggregateType=n,this.Constructor=t,this.snapshots=a,this.subscriptions=[]}return _createClass(u,[{key:"subscribe",value:function(e){this.subscriptions.push(e)}}]),u}();exports.default=Repository;
+"use strict";Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=void 0;var _v=_interopRequireDefault(require("uuid/v4"));function _interopRequireDefault(e){return e&&e.__esModule?e:{default:e}}function _objectSpread(r){for(var e=1;e<arguments.length;e++){var t=null!=arguments[e]?arguments[e]:{},n=Object.keys(t);"function"==typeof Object.getOwnPropertySymbols&&(n=n.concat(Object.getOwnPropertySymbols(t).filter(function(e){return Object.getOwnPropertyDescriptor(t,e).enumerable}))),n.forEach(function(e){_defineProperty(r,e,t[e])})}return r}function asyncGeneratorStep(e,r,t,n,o,a,u){try{var s=e[a](u),i=s.value}catch(e){return void t(e)}s.done?r(i):Promise.resolve(i).then(n,o)}function _asyncToGenerator(s){return function(){var e=this,u=arguments;return new Promise(function(r,t){var n=s.apply(e,u);function o(e){asyncGeneratorStep(n,r,t,o,a,"next",e)}function a(e){asyncGeneratorStep(n,r,t,o,a,"throw",e)}o(void 0)})}}function _classCallCheck(e,r){if(!(e instanceof r))throw new TypeError("Cannot call a class as a function")}function _defineProperties(e,r){for(var t=0;t<r.length;t++){var n=r[t];n.enumerable=n.enumerable||!1,n.configurable=!0,"value"in n&&(n.writable=!0),Object.defineProperty(e,n.key,n)}}function _createClass(e,r,t){return r&&_defineProperties(e.prototype,r),t&&_defineProperties(e,t),e}function _defineProperty(e,r,t){return r in e?Object.defineProperty(e,r,{value:t,enumerable:!0,configurable:!0,writable:!0}):e[r]=t,e}var Repository=function(){function u(e){var s=this,r=e.eventStore,t=e.constructor,n=e.aggregateType,o=e.snapshots,a=void 0===o||o;_classCallCheck(this,u),_defineProperty(this,"get",function(){var r=_asyncToGenerator(regeneratorRuntime.mark(function e(r){var t,n,o,a,u;return regeneratorRuntime.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,s.eventStore.getEvents({aggregateType:s.aggregateType,aggregateId:r});case 2:if(n=e.sent,o=n.events,(a=n.snapshot)||o.length){e.next=7;break}return e.abrupt("return",null);case 7:if(u=new s.Constructor({id:r,events:o,snapshot:null==a?void 0:a.body,version:null==a?void 0:a.version}),s.snapshots&&((null===(t=u.$snapshotSchedule)||void 0===t?void 0:t.every)||Number.MAX_VALUE)<=o.length)return e.next=11,s.eventStore.saveSnapshot(s.aggregateType,u);e.next=11;break;case 11:return e.abrupt("return",u);case 12:case"end":return e.stop()}},e,this)}));return function(e){return r.apply(this,arguments)}}()),_defineProperty(this,"create",function(e){return new s.Constructor({id:e||(0,_v.default)()})}),_defineProperty(this,"record",function(){var r=_asyncToGenerator(regeneratorRuntime.mark(function e(r){var t;return regeneratorRuntime.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,s.eventStore.record(r.map(function(e){return _objectSpread({},e,{aggregate:s.aggregateType})}));case 2:return t=e.sent,e.next=5,s.notifySubscribers(t);case 5:case"end":return e.stop()}},e,this)}));return function(e){return r.apply(this,arguments)}}()),_defineProperty(this,"notifySubscribers",function(){var r=_asyncToGenerator(regeneratorRuntime.mark(function e(t){return regeneratorRuntime.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,Promise.all(s.subscriptions.map(function(){var r=_asyncToGenerator(regeneratorRuntime.mark(function e(r){return regeneratorRuntime.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return e.prev=0,e.next=3,r(t);case 3:return e.abrupt("return",e.sent);case 6:return e.prev=6,e.t0=e.catch(0),console.log(e.t0),e.abrupt("return",null);case 10:case"end":return e.stop()}},e,this,[[0,6]])}));return function(e){return r.apply(this,arguments)}}()));case 2:return e.abrupt("return",e.sent);case 3:case"end":return e.stop()}},e,this)}));return function(e){return r.apply(this,arguments)}}()),_defineProperty(this,"getEvents",function(){var r=_asyncToGenerator(regeneratorRuntime.mark(function e(r){return regeneratorRuntime.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,s.eventStore.getEvents({aggregateType:s.aggregateType,aggregateId:r});case 2:return e.abrupt("return",e.sent);case 3:case"end":return e.stop()}},e,this)}));return function(e){return r.apply(this,arguments)}}()),this.eventStore=r,this.aggregateType=n,this.Constructor=t,this.snapshots=a,this.subscriptions=[]}return _createClass(u,[{key:"subscribe",value:function(e){this.subscriptions.push(e)}}]),u}();exports.default=Repository;
 //# sourceMappingURL=maps/repository.js.map
