@@ -1,2 +1,2 @@
-"use strict";Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=check;var _authorizationError=_interopRequireDefault(require("./authorizationError"));function _interopRequireDefault(t){return t&&t.__esModule?t:{default:t}}function asyncGeneratorStep(t,e,r,n,o,u,c){try{var i=t[u](c),a=i.value}catch(t){return void r(t)}i.done?e(a):Promise.resolve(a).then(n,o)}function _asyncToGenerator(i){return function(){var t=this,c=arguments;return new Promise(function(e,r){var n=i.apply(t,c);function o(t){asyncGeneratorStep(n,e,r,o,u,"next",t)}function u(t){asyncGeneratorStep(n,e,r,o,u,"throw",t)}o(void 0)})}}function isNativeReflectConstruct(){if("undefined"==typeof Reflect||!Reflect.construct)return!1;if(Reflect.construct.sham)return!1;if("function"==typeof Proxy)return!0;try{return Date.prototype.toString.call(Reflect.construct(Date,[],function(){})),!0}catch(t){return!1}}function _construct(t,e,r){return(_construct=isNativeReflectConstruct()?Reflect.construct:function(t,e,r){var n=[null];n.push.apply(n,e);var o=new(Function.bind.apply(t,n));return r&&_setPrototypeOf(o,r.prototype),o}).apply(null,arguments)}function _setPrototypeOf(t,e){return(_setPrototypeOf=Object.setPrototypeOf||function(t,e){return t.__proto__=e,t})(t,e)}function check(){var t=0<arguments.length&&void 0!==arguments[0]?arguments[0]:{},e=t.rights,c=void 0===e?[]:e,r=t.message,i=void 0===r?"User is not allowed to issue this command":r;return function(u,t,e){return function(){for(var t=arguments.length,e=new Array(t),r=0;r<t;r++)e[r]=arguments[r];var n=_construct(u,e),o=n.authorize;return n.authorize=function(){var r=_asyncToGenerator(regeneratorRuntime.mark(function t(e,r){return regeneratorRuntime.wrap(function(t){for(;;)switch(t.prev=t.next){case 0:if(c.some(function(t){return e.$identity.rights.includes(t)})){t.next=2;break}throw new _authorizationError.default({aggregate:r,command:e,message:i});case 2:if(!o){t.next=5;break}return t.next=5,o(e,r);case 5:case"end":return t.stop()}},t,this)}));return function(t,e){return r.apply(this,arguments)}}(),n}}}
+"use strict";Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=check;var _authorizationError=_interopRequireDefault(require("./authorizationError"));function _interopRequireDefault(t){return t&&t.__esModule?t:{default:t}}function asyncGeneratorStep(t,e,r,n,o,u,c){try{var i=t[u](c),a=i.value}catch(t){return void r(t)}i.done?e(a):Promise.resolve(a).then(n,o)}function _asyncToGenerator(i){return function(){var t=this,c=arguments;return new Promise(function(e,r){var n=i.apply(t,c);function o(t){asyncGeneratorStep(n,e,r,o,u,"next",t)}function u(t){asyncGeneratorStep(n,e,r,o,u,"throw",t)}o(void 0)})}}function isNativeReflectConstruct(){if("undefined"==typeof Reflect||!Reflect.construct)return!1;if(Reflect.construct.sham)return!1;if("function"==typeof Proxy)return!0;try{return Date.prototype.toString.call(Reflect.construct(Date,[],function(){})),!0}catch(t){return!1}}function _construct(t,e,r){return(_construct=isNativeReflectConstruct()?Reflect.construct:function(t,e,r){var n=[null];n.push.apply(n,e);var o=new(Function.bind.apply(t,n));return r&&_setPrototypeOf(o,r.prototype),o}).apply(null,arguments)}function _setPrototypeOf(t,e){return(_setPrototypeOf=Object.setPrototypeOf||function(t,e){return t.__proto__=e,t})(t,e)}function check(){var t=0<arguments.length&&void 0!==arguments[0]?arguments[0]:{},e=t.rights,c=void 0===e?[]:e,r=t.message,i=void 0===r?"User is not allowed to issue this command":r;return function(u,t,e){return function(){for(var t=arguments.length,e=new Array(t),r=0;r<t;r++)e[r]=arguments[r];var n=_construct(u,e),o=n.authorize;return n.authorize=function(){var r=_asyncToGenerator(regeneratorRuntime.mark(function t(e,r){return regeneratorRuntime.wrap(function(t){for(;;)switch(t.prev=t.next){case 0:if(c.some(function(t){return e.$identity.rights.includes(t)})){t.next=2;break}throw new _authorizationError.default({aggregate:r,command:e,message:i});case 2:if(o)return t.next=5,o(e,r);t.next=5;break;case 5:case"end":return t.stop()}},t,this)}));return function(t,e){return r.apply(this,arguments)}}(),n}}}
 //# sourceMappingURL=../maps/auth/authorizeDecorator.js.map
