@@ -1,2 +1,2 @@
-"use strict";Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=void 0;var _identity=_interopRequireDefault(require("../auth/identity")),_schedule=_interopRequireDefault(require("./schedule")),_commandFailure=_interopRequireWildcard(require("./commandFailure"));function _interopRequireWildcard(e){if(e&&e.__esModule)return e;var r={};if(null!=e)for(var t in e)if(Object.prototype.hasOwnProperty.call(e,t)){var n=Object.defineProperty&&Object.getOwnPropertyDescriptor?Object.getOwnPropertyDescriptor(e,t):{};n.get||n.set?Object.defineProperty(r,t,n):r[t]=e[t]}return r.default=e,r}function _interopRequireDefault(e){return e&&e.__esModule?e:{default:e}}function asyncGeneratorStep(e,r,t,n,a,i,u){try{var c=e[i](u),o=c.value}catch(e){return void t(e)}c.done?r(o):Promise.resolve(o).then(n,a)}function _asyncToGenerator(c){return function(){var e=this,u=arguments;return new Promise(function(r,t){var n=c.apply(e,u);function a(e){asyncGeneratorStep(n,r,t,a,i,"next",e)}function i(e){asyncGeneratorStep(n,r,t,a,i,"throw",e)}a(void 0)})}}function _classCallCheck(e,r){if(!(e instanceof r))throw new TypeError("Cannot call a class as a function")}function _defineProperty(e,r,t){return r in e?Object.defineProperty(e,r,{value:t,enumerable:!0,configurable:!0,writable:!0}):e[r]=t,e}var RabbitScheduler=function e(r){var c=this,i=r.channel,t=r.clock,n=r.deliverer;_classCallCheck(this,e),_defineProperty(this,"schedule",function(){var r=_asyncToGenerator(regeneratorRuntime.mark(function e(r){var t,n,a,i,u;return regeneratorRuntime.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return t=r.service,n=r.target,a=r.clock,i=r.due,(u=r.command).$identity=_identity.default.system,u.$scheduler=new _schedule.default({service:t,target:n,due:i,clock:a||c.clock}),e.next=5,c.publish(JSON.stringify(u));case 5:case"end":return e.stop()}},e,this)}));return function(e){return r.apply(this,arguments)}}()),_defineProperty(this,"execute",function(){var r=_asyncToGenerator(regeneratorRuntime.mark(function e(r){var t,n,a;return regeneratorRuntime.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return t=r.service,n=r.target,(a=r.command).$identity=_identity.default.system,a.$scheduler=new _schedule.default({service:t,target:n}),e.next=5,c.deliver(a);case 5:case"end":return e.stop()}},e,this)}));return function(e){return r.apply(this,arguments)}}()),_defineProperty(this,"deliver",function(){var r=_asyncToGenerator(regeneratorRuntime.mark(function e(n){return regeneratorRuntime.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return e.abrupt("return",new Promise(function(){var t=_asyncToGenerator(regeneratorRuntime.mark(function e(r,t){return regeneratorRuntime.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return e.prev=0,n.$scheduler.due=null,e.next=4,c.deliverer.deliver({service:n.$scheduler.service,target:n.$scheduler.target,command:n});case 4:r(),e.next=10;break;case 7:e.prev=7,e.t0=e.catch(0),t(e.t0);case 10:case"end":return e.stop()}},e,this,[[0,7]])}));return function(e,r){return t.apply(this,arguments)}}()));case 1:case"end":return e.stop()}},e,this)}));return function(e){return r.apply(this,arguments)}}()),this.channel=i,this.clock=t,this.deliverer=n;var a="scheduledCommands";i.assertQueue(a,{durable:!0}).then(function(){return i.consume(a,function(){var r=_asyncToGenerator(regeneratorRuntime.mark(function e(n){var a;return regeneratorRuntime.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return a=JSON.parse(n.content),e.next=3,c.deliver(a).then(function(){i.ack(n)}).catch(function(){var r=_asyncToGenerator(regeneratorRuntime.mark(function e(r){var t;return regeneratorRuntime.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:if(t=new _commandFailure.default({error:r,command:a}),!r.handler){e.next=4;break}return e.next=4,r.handler.handleDeliveryError(t,r.aggregate);case 4:i.nack(n,!1,t.failureAction instanceof _commandFailure.Retry&&!n.fields.redelivered);case 5:case"end":return e.stop()}},e,this)}));return function(e){return r.apply(this,arguments)}}());case 3:case"end":return e.stop()}},e,this)}));return function(e){return r.apply(this,arguments)}}())}),this.publish=function(){var r=_asyncToGenerator(regeneratorRuntime.mark(function e(r){return regeneratorRuntime.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,i.sendToQueue(a,new Buffer(r));case 2:case"end":return e.stop()}},e,this)}));return function(e){return r.apply(this,arguments)}}()};exports.default=RabbitScheduler;
+"use strict";Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=void 0;var _identity=_interopRequireDefault(require("../auth/identity")),_schedule=_interopRequireDefault(require("./schedule")),_commandFailure=_interopRequireWildcard(require("./commandFailure"));function _interopRequireWildcard(e){if(e&&e.__esModule)return e;var r={};if(null!=e)for(var t in e)if(Object.prototype.hasOwnProperty.call(e,t)){var n=Object.defineProperty&&Object.getOwnPropertyDescriptor?Object.getOwnPropertyDescriptor(e,t):{};n.get||n.set?Object.defineProperty(r,t,n):r[t]=e[t]}return r.default=e,r}function _interopRequireDefault(e){return e&&e.__esModule?e:{default:e}}function asyncGeneratorStep(e,r,t,n,a,i,u){try{var c=e[i](u),o=c.value}catch(e){return void t(e)}c.done?r(o):Promise.resolve(o).then(n,a)}function _asyncToGenerator(c){return function(){var e=this,u=arguments;return new Promise(function(r,t){var n=c.apply(e,u);function a(e){asyncGeneratorStep(n,r,t,a,i,"next",e)}function i(e){asyncGeneratorStep(n,r,t,a,i,"throw",e)}a(void 0)})}}function _classCallCheck(e,r){if(!(e instanceof r))throw new TypeError("Cannot call a class as a function")}function _defineProperty(e,r,t){return r in e?Object.defineProperty(e,r,{value:t,enumerable:!0,configurable:!0,writable:!0}):e[r]=t,e}var RabbitScheduler=function e(r){var c=this,i=r.channel,t=r.clock,n=r.deliverer;_classCallCheck(this,e),_defineProperty(this,"schedule",function(){var r=_asyncToGenerator(regeneratorRuntime.mark(function e(r){var t,n,a,i,u;return regeneratorRuntime.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return t=r.service,n=r.target,a=r.clock,i=r.due,(u=r.command).$identity=_identity.default.system,u.$scheduler=new _schedule.default({service:t,target:n,due:i,clock:a||c.clock}),e.next=5,c.publish(JSON.stringify(u));case 5:case"end":return e.stop()}},e,this)}));return function(e){return r.apply(this,arguments)}}()),_defineProperty(this,"execute",function(){var r=_asyncToGenerator(regeneratorRuntime.mark(function e(r){var t,n,a;return regeneratorRuntime.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return t=r.service,n=r.target,(a=r.command).$identity=_identity.default.system,a.$scheduler=new _schedule.default({service:t,target:n}),e.next=5,c.deliver(a);case 5:case"end":return e.stop()}},e,this)}));return function(e){return r.apply(this,arguments)}}()),_defineProperty(this,"deliver",function(){var r=_asyncToGenerator(regeneratorRuntime.mark(function e(n){return regeneratorRuntime.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return e.abrupt("return",new Promise(function(){var t=_asyncToGenerator(regeneratorRuntime.mark(function e(r,t){return regeneratorRuntime.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return e.prev=0,n.$scheduler.due=null,e.next=4,c.deliverer.deliver({service:n.$scheduler.service,target:n.$scheduler.target,command:n});case 4:r(),e.next=10;break;case 7:e.prev=7,e.t0=e.catch(0),t(e.t0);case 10:case"end":return e.stop()}},e,this,[[0,7]])}));return function(e,r){return t.apply(this,arguments)}}()));case 1:case"end":return e.stop()}},e,this)}));return function(e){return r.apply(this,arguments)}}()),this.channel=i,this.clock=t,this.deliverer=n;var a="scheduledCommands";i.assertQueue(a,{durable:!0}).then(function(){return i.consume(a,function(){var r=_asyncToGenerator(regeneratorRuntime.mark(function e(n){var a;return regeneratorRuntime.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return a=JSON.parse(n.content),e.next=3,c.deliver(a).then(function(){i.ack(n)}).catch(function(){var r=_asyncToGenerator(regeneratorRuntime.mark(function e(r){var t;return regeneratorRuntime.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:if(t=new _commandFailure.default({error:r,command:a}),r.handler)return e.next=4,r.handler.handleDeliveryError(t,r.aggregate);e.next=4;break;case 4:i.nack(n,!1,t.failureAction instanceof _commandFailure.Retry&&!n.fields.redelivered);case 5:case"end":return e.stop()}},e,this)}));return function(e){return r.apply(this,arguments)}}());case 3:case"end":return e.stop()}},e,this)}));return function(e){return r.apply(this,arguments)}}())}),this.publish=function(){var r=_asyncToGenerator(regeneratorRuntime.mark(function e(r){return regeneratorRuntime.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,i.sendToQueue(a,new Buffer(r));case 2:case"end":return e.stop()}},e,this)}));return function(e){return r.apply(this,arguments)}}()};exports.default=RabbitScheduler;
 //# sourceMappingURL=../maps/commandScheduling/rabbitScheduler.js.map
